requirements-runtime.txt
gradio>=4.0.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.25.0
tqdm>=4.60.0
